feat: refactorizar la obtención de datos de usuario y el manejo de errores, mejorar la retroalimentación de la interfaz de usuario y eliminar el código no utilizado
